requirements.txt
torch
transformers
accelerate
qwen-vl-utils
pymupdf
pillow
json_repair
sentence-transformers
tqdm
